This simplifies the gallery, preview, and main component. It also allows users to keep track of (and cancel!) pending requests. It's also a step towards potentially supporting Img2Img as well, since the `params` stored in `ongoing` and `requests` can just be replaced with a `Text of Txt2Img.Query.t | Img of Img2Img.Query.t` union type.
The biggest missing piece (other than styling) is if there is already a request running when the app is started; e.g. if the page was refreshed with requests still processing / in the queue. This can be handled by adding an "on activate" effect to `request_queue` that blocks the queue dispatcher, checks for progress, and if anything is found, sets `ongoing` to "`Preoccupied", which should reflect in the preview accordingly.

Signed-off-by: Alexander Skvortsov <sasha.skvortsov109@gmail.com>